DRAFT/WIP Fix incoming file transfers
With TOOD the concept if IQ request handlers was introduced in
Smack. This doesn't allow packet collectors/listeners to filter for
incoming IQ requests. Unfortunately the file transfer code relied on
this being able, so it broke with the change.

There were two places where the file transfer code was listening for
incoming IQ requests:
- InitationListener(s)
- Negotiator(s)

With this change, we let the InitiationListener signal the existence of
an incoming initation request, send by an IQ of type 'set', using the
newly created EventManager utility.

The negotiator waits for those events to arrive and proceedes as it would
have done when the packet collector was used.